# enterprise-vision-mlops/contracts/control-panel/control-panel.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise Vision MLOps Control Panel API",
    "version": "2026-07-12.post-w7.multi-cycle-release.v1",
    "description": "Multi-cycle metadata and guarded mutation contract for the enterprise MLOps Control Panel, including runtime task dispatch, CI evidence, deployment intent state transitions, organizational service scope, drift review, CD/CT gates, and real-test Torch EfficientNet model matrices."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "local FastAPI service"
    }
  ],
  "paths": {
    "/control-panel/v1/cycles": {
      "get": {
        "summary": "Search and list live and historical lifecycle cycles.",
        "operationId": "listCycles",
        "parameters": [
          {"name": "status", "in": "query", "required": false, "schema": {"$ref": "#/components/schemas/State"}},
          {"name": "environment", "in": "query", "required": false, "schema": {"type": "string", "enum": ["dev", "test", "staging", "pre-production", "production"]}},
          {"name": "query", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 200, "default": 50}}
        ],
        "responses": {
          "200": {
            "description": "cycle catalog",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CycleRunList"}}}
          }
        }
      }
    },
    "/control-panel/v1/cycles/{cycle_id}": {
      "get": {
        "summary": "Return one lifecycle cycle with lineage, stages, artifacts, serving state, and resource refs.",
        "operationId": "getCycle",
        "parameters": [
          {
            "name": "cycle_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "cycle detail",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CycleRun"}
              }
            }
          },
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
    "/control-panel/v1/resources": {
      "get": {
        "summary": "Return normalized Kubernetes and external worker resource state.",
        "operationId": "listResources",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "required": false,
            "schema": {"type": "string"}
          },
          {
            "name": "owner_issue",
            "in": "query",
            "required": false,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "resource list",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/RuntimeResourceList"}
              }
            }
          }
        }
      }
    },
...
    "/control-panel/v1/promotion-policy/evaluate": {
      "post": {
        "summary": "Evaluate target environment promotion policy from current server evidence.",
        "operationId": "evaluatePromotionPolicy",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/PromotionPolicyRequest"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "server-computed promotion policy decision",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/PromotionPolicyDecision"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"}
        }
      }
    },
    "/control-panel/v1/deployment-intents": {
      "get": {
        "summary": "Return the auditable deployment intent ledger.",
        "operationId": "listDeploymentIntents",
        "parameters": [
          {"name": "cycle_id", "in": "query", "required": false, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "deployment intent ledger",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeploymentIntentList"}}}
          }
        }
      },
      "post": {
        "summary": "Admit a dry-run deployment intent after CI, readiness, and environment policy validation.",
        "operationId": "createDeploymentIntent",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeploymentIntentRequest"}}}
        },
        "responses": {
          "202": {
            "description": "dry-run deployment intent admitted",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeploymentIntent"}}}
          },
          "409": {"description": "CI, readiness, policy, or ledger admission blocked"}
        }
      }
    },
    "/control-panel/v1/deployment-intents/{intent_id}/request-approval": {
      "post": {
        "summary": "Move a dry-run deployment intent to pending approval.",
        "operationId": "requestDeploymentApproval",
        "parameters": [{"name": "intent_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeploymentTransitionRequest"}}}},
        "responses": {
          "202": {"description": "approval requested", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeploymentIntent"}}}},
          "409": {"description": "transition rejected"}
        }
      }
    },
    "/control-panel/v1/deployment-intents/{intent_id}/approve": {
      "post": {
        "summary": "Record a separate authorized approver without applying the deployment.",
        "operationId": "approveDeploymentIntent",
        "parameters": [{"name": "intent_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeploymentTransitionRequest"}}}},
        "responses": {
          "202": {"description": "approval recorded", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeploymentIntent"}}}},
          "409": {"description": "approval or separation-of-duties check rejected"}
        }
      }
    },
    "/control-panel/v1/deployment-intents/{intent_id}/queue": {
      "post": {
        "summary": "Revalidate all gates and queue an approved intent for the isolated executor.",
        "operationId": "queueDeploymentIntent",
        "parameters": [{"name": "intent_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeploymentTransitionRequest"}}}},
        "responses": {
          "202": {"description": "deployment intent queued", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeploymentIntent"}}}},
          "409": {"description": "current CI, readiness, policy, or version check rejected"}
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "invalid request",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      },
      "NotFound": {
        "description": "resource not found",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      }
    },
    "schemas": {
      "CycleRun": {
        "type": "object",
        "required": ["cycle_id", "status", "started_at", "stages", "dataset", "model", "serving", "resources"],
        "properties": {
          "cycle_id": {"type": "string"},
          "status": {"$ref": "#/components/schemas/State"},
          "started_at": {"type": "string", "format": "date-time"},
          "finished_at": {"type": ["string", "null"], "format": "date-time"},
          "owner_issue": {"type": "string"},
          "tenant": {"$ref": "#/components/schemas/OrgContext"},
          "environment": {"$ref": "#/components/schemas/EnvironmentRef"},
          "airflow": {"$ref": "#/components/schemas/AirflowRef"},
          "mlflow": {"$ref": "#/components/schemas/MLflowRef"},
          "data_pipeline": {"$ref": "#/components/schemas/DataPipelineReadiness"},
          "experiment_pipeline": {"$ref": "#/components/schemas/ExperimentPipelineReadiness"},
          "readiness_evaluation": {"$ref": "#/components/schemas/ArtifactReadinessEvaluation"},
          "promotion_policy": {"$ref": "#/components/schemas/PromotionPolicyDecision"},
          "ci_evidence": {"anyOf": [{"$ref": "#/components/schemas/CIEvidenceValidation"}, {"type": "null"}]},
          "latest_deployment_intent": {"anyOf": [{"$ref": "#/components/schemas/DeploymentIntent"}, {"type": "null"}]},
          "dataset": {"$ref": "#/components/schemas/DatasetVersion"},
          "model": {"$ref": "#/components/schemas/ModelVersion"},
          "model_matrix": {"$ref": "#/components/schemas/ModelExperimentMatrix"},
          "metrics": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/Metric"}
          },
          "promotion_gate": {"$ref": "#/components/schemas/PromotionGate"},
          "drift": {"$ref": "#/components/schemas/DriftState"},
          "cdct_gate": {"$ref": "#/components/schemas/CDCTGate"},
          "serving": {"$ref": "#/components/schemas/ServingState"},
          "stages": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/PipelineStage"}
          },
          "resources": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ResourceRef"}
          },
          "artifacts": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ArtifactRef"}
          }
        }
      },
      "CycleRunSummary": {
        "type": "object",
        "required": ["cycle_id", "status", "started_at", "dataset_id", "dataset_version", "model_name", "model_version", "model_stage", "owner_issue", "stage_count", "progress", "live"],
        "properties": {
          "cycle_id": {"type": "string"},
          "status": {"$ref": "#/components/schemas/State"},
          "started_at": {"type": "string", "format": "date-time"},
          "finished_at": {"type": ["string", "null"], "format": "date-time"},
          "dataset_id": {"type": "string"},
          "dataset_version": {"type": "string"},
          "model_name": {"type": "string"},
          "model_version": {"type": "string"},
          "model_stage": {"type": "string"},
          "environment": {"type": ["string", "null"], "enum": ["dev", "test", "staging", "pre-production", "production", null]},
          "owner_issue": {"type": "string"},
          "stage_count": {"type": "integer", "minimum": 0},
          "progress": {"type": "number", "minimum": 0, "maximum": 1},
          "source_uri": {"type": ["string", "null"]},
          "live": {"type": "boolean"}
        }
      },
      "CycleRunList": {
        "type": "object",
        "required": ["cycles", "latest_cycle_id", "total"],
        "properties": {
          "cycles": {"type": "array", "items": {"$ref": "#/components/schemas/CycleRunSummary"}},
          "latest_cycle_id": {"type": "string"},
          "selected_cycle_id": {"type": ["string", "null"]},
          "total": {"type": "integer", "minimum": 0}
        }
      },
...
      "DeploymentIntentRequest": {
        "type": "object",
        "required": ["target_environment", "target_namespace", "target", "actor", "reason", "dry_run"],
        "properties": {
          "cycle_id": {"type": ["string", "null"]},
          "target_environment": {"type": "string", "enum": ["dev", "test", "staging", "pre-production", "production"]},
          "target_namespace": {"type": "string"},
          "target": {"$ref": "#/components/schemas/ResourceRef"},
          "actor": {"type": "string"},
          "reason": {"type": "string"},
          "dry_run": {"type": "boolean", "const": true}
        }
      },
      "DeploymentTransitionRequest": {
        "type": "object",
        "required": ["actor", "reason", "expected_version"],
        "properties": {
          "actor": {"type": "string"},
          "reason": {"type": "string"},
          "expected_version": {"type": "integer", "minimum": 1}
        }
      },
...
      "DeploymentIntent": {
        "allOf": [
          {"$ref": "#/components/schemas/DeploymentIntentRequest"},
          {
            "type": "object",
            "required": ["intent_id", "state", "version", "created_at", "updated_at", "ci_evidence", "ci_evidence_uri", "ci_bundle_digest", "readiness_evaluation_id", "promotion_policy", "model_candidate_id", "model_artifact_uri", "model_digest", "image_digest", "config_render_digest", "rollback_reference", "manifest_ref", "transitions"],
            "properties": {
              "intent_id": {"type": "string"},
              "state": {"$ref": "#/components/schemas/DeploymentIntentState"},
              "version": {"type": "integer", "minimum": 1},
              "created_at": {"type": "string", "format": "date-time"},
              "updated_at": {"type": "string", "format": "date-time"},
              "ci_evidence": {"$ref": "#/components/schemas/CIEvidenceValidation"},
              "ci_evidence_uri": {"type": "string"},
              "ci_bundle_digest": {"type": "string"},
              "readiness_evaluation_id": {"type": "string"},
              "promotion_policy": {"$ref": "#/components/schemas/PromotionPolicyDecision"},
              "model_candidate_id": {"type": "string"},
              "model_artifact_uri": {"type": "string"},
              "model_digest": {"type": "string"},
              "image_digest": {"type": "string"},
              "config_render_digest": {"type": "string"},
              "rollback_reference": {"type": "string"},
              "manifest_ref": {"type": "string"},
              "audit_uri": {"type": ["string", "null"]},
              "approver": {"type": ["string", "null"]},
              "approved_at": {"type": ["string", "null"], "format": "date-time"},
              "transitions": {"type": "array", "items": {"$ref": "#/components/schemas/DeploymentTransition"}},
              "execution_result": {"anyOf": [{"$ref": "#/components/schemas/DeploymentExecutionResult"}, {"type": "null"}]}
            }
          }
        ]
      },
      "DeploymentIntentList": {
        "type": "object",
        "required": ["intents", "status", "blockers"],
        "properties": {
          "intents": {"type": "array", "items": {"$ref": "#/components/schemas/DeploymentIntent"}},
          "status": {"$ref": "#/components/schemas/State"},
          "blockers": {"type": "array", "items": {"type": "string"}}
        }
      },
...
      "OrgContext": {
        "type": "object",
        "required": ["team_id", "department", "service_scope"],
        "properties": {
          "team_id": {"type": "string"},
          "department": {"type": "string"},
          "product_area": {"type": ["string", "null"]},
          "service_scope": {"type": "string", "enum": ["internal-team", "internal-department", "external-production"]},
          "data_owner": {"type": ["string", "null"]},
          "model_owner": {"type": ["string", "null"]},
          "ops_owner": {"type": ["string", "null"]},
          "ownership_status": {"$ref": "#/components/schemas/State"},
          "missing_owners": {
            "type": "array",
            "items": {"type": "string"}
          }
        }
      },
      "EnvironmentRef": {
        "type": "object",
        "required": ["name", "tier", "promotion_state"],
        "properties": {
          "name": {"type": "string"},
          "tier": {"type": "string", "enum": ["dev", "test", "staging", "pre-production", "production"]},
          "promotion_state": {"type": "string", "enum": ["draft", "candidate", "approved", "blocked", "deployed", "rolled_back"]},
          "cluster": {"type": ["string", "null"]},
          "namespace": {"type": ["string", "null"]},
          "release_ref": {"type": ["string", "null"]},
          "approval_policy": {"type": ["string", "null"]},
          "promotion_blockers": {
            "type": "array",
            "items": {"type": "string"}
          }
        }
      },
      "PromotionPolicyRequest": {
        "type": "object",
        "required": ["target_environment", "target_namespace", "requester"],
        "properties": {
          "target_environment": {
            "type": "string",
            "enum": ["dev", "test", "staging", "pre-production", "production"]
          },
          "target_namespace": {"type": "string"},
          "requester": {"type": "string"},
          "approver": {"type": ["string", "null"]}
        }
      },
...
      "PromotionPolicyDecision": {
        "type": "object",
        "required": [
          "decision_id",
          "policy_version",
          "decision",
          "status",
          "target_environment",
          "target_namespace",
          "requester",
          "approval_policy",
          "evaluated_at",
          "input_digest"
        ],
        "properties": {
          "schema_version": {"type": "string"},
          "decision_id": {"type": "string"},
          "policy_version": {"type": "string"},
          "decision": {
            "type": "string",
            "enum": ["allow", "pending_approval", "blocked"]
          },
          "status": {"$ref": "#/components/schemas/State"},
          "target_environment": {
            "type": "string",
            "enum": ["dev", "test", "staging", "pre-production", "production"]
          },
          "target_namespace": {"type": "string"},
          "requester": {"type": "string"},
          "approver": {"type": ["string", "null"]},
          "approval_policy": {"type": "string"},
          "evaluated_at": {"type": "string", "format": "date-time"},
          "input_digest": {"type": "string"},
          "required_checks": {
            "type": "array",
            "items": {"type": "string"}
          },
          "required_approvals": {
            "type": "array",
            "items": {"type": "string"}
          },
          "reason_codes": {
            "type": "array",
            "items": {"type": "string"}
          },
          "checks": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/PromotionPolicyCheck"}
          },
          "audit_uri": {"type": ["string", "null"]}
        }
      },
      "DataPipelineReadiness": {
        "type": "object",
        "required": ["contract_status", "quality_status", "lineage_status", "replay_ready"],
        "properties": {
          "contract_status": {"$ref": "#/components/schemas/State"},
          "quality_status": {"$ref": "#/components/schemas/State"},
          "lineage_status": {"$ref": "#/components/schemas/State"},
          "replay_ready": {"type": "boolean"},
          "source_policy_uri": {"type": ["string", "null"]},
          "quality_report_uri": {"type": ["string", "null"]},
          "lineage_uri": {"type": ["string", "null"]},
          "backfill_window": {"type": ["string", "null"]},
          "owner_approval_required": {"type": "boolean"},
          "owner_approval_status": {"$ref": "#/components/schemas/State"},
          "owner_approval_actor": {"type": ["string", "null"]},
          "blockers": {
            "type": "array",
            "items": {"type": "string"}
          }
        }
      },
...
      "ModelVersion": {
        "type": "object",
        "required": ["model_name", "version", "stage", "model_type", "registry_uri"],
        "properties": {
          "model_name": {"type": "string"},
          "version": {"type": "string"},
          "stage": {"type": "string"},
          "model_type": {"type": "string"},
          "registry_uri": {"type": "string"},
          "source_run_id": {"type": "string"},
          "dataset_version": {"type": "string"}
        }
      },
...
      "RuntimeResourceList": {
        "type": "object",
        "required": ["resources", "observation_status"],
        "properties": {
          "resources": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/RuntimeResource"}
          },
          "observation_status": {
            "type": "string",
            "enum": ["live", "stale", "projected", "unavailable"]
          },
          "observed_at": {"type": ["string", "null"], "format": "date-time"},
          "snapshot_age_seconds": {"type": ["number", "null"], "minimum": 0},
          "cluster_context": {"type": ["string", "null"]},
          "snapshot_uri": {"type": ["string", "null"]},
          "observation_message": {"type": ["string", "null"]}
        }
      },
...
      "AirflowRef": {
        "type": "object",
        "properties": {
          "mode": {"type": "string", "enum": ["external-compose", "in-cluster", "managed-service"]},
          "control_mode": {"type": "string", "enum": ["read-only", "rest-api", "cli-bridge", "kubernetes-operator"]},
          "dag_id": {"type": "string"},
          "dag_run_id": {"type": "string"},
          "task_id": {"type": ["string", "null"]},
          "namespace": {"type": ["string", "null"]},
          "deployment_name": {"type": ["string", "null"]},
          "contract_config_map": {"type": ["string", "null"]},
          "connection_status": {"$ref": "#/components/schemas/State"},
          "url": {"type": ["string", "null"]}
        }
      },
...
      "MLflowRef": {
        "type": "object",
        "properties": {
          "experiment_id": {"type": "string"},
          "run_id": {"type": "string"},
          "model_name": {"type": "string"},
          "model_version": {"type": "string"},
          "url": {"type": ["string", "null"]}
        }
      },
...
      "State": {
        "type": "string",
        "enum": ["unknown", "queued", "running", "pass", "warn", "fail", "blocked", "done", "cancelled"]
      },
...
    }
  }
}